Refactor to machinery-based use case
